requirements.txt
opencv-python>=4.8.0.74
pyzbar>=0.1.9
PyQt5>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0 